FHIR client connection

FHIR client connection

Client(
  app_id = NULL,
  app_secret = NULL,
  api_base,
  redirect_uri = NULL,
  patient_id = NULL,
  scope = list(),
  launch_token = NULL,
  state = NULL
)

Arguments

app_id
The app ID.

app_secret
The client secret.

api_base
The FHIR service to connect.

redirect_uri
The callback/redirect URL for your app.

patient_id
The patient id against which to operate, if already known.

scope
Space-separated list of scopes to request, if other than default.

launch_token
The launch token.

state
The client state.

Examples

settings = list('app_id'= 'my_web_app', 'api_base'= 'https://r4.smarthealthit.org')
smart <- Client("my_app", api_base = "https://r4.smarthealthit.org")